refactor: Update to PublicLayout file
